version lista

+++ src/main/java/com/miportafolio/ms1/repositories/UsuarioRolRepository.java
@@ -3,5 +3,5 @@ package com.miportafolio.ms1.repositories;
 import com.miportafolio.ms1.models.UsuarioRol;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UsuarioRolRepository extends JpaRepository <UsuarioRol, Integer> {
+public interface UsuarioRolRepository extends JpaRepository<UsuarioRol, Long> {
 }
